servicebus: moving the Topic Name validate func
