+++ azurerm/internal/services/servicebus/servicebus_subscription_data_source.go
@@ -37,7 +37,7 @@ func dataSourceArmServiceBusSubscription() *schema.Resource {
 			"topic_name": {
 				Type:         schema.TypeString,
 				Required:     true,
-				ValidateFunc: azure.ValidateServiceBusTopicName(),
+				ValidateFunc: validate.TopicName(),
 			},
 
 			"auto_delete_on_idle": {
