+++ azurerm/internal/services/servicebus/servicebus_topic_authorization_rule_resource.go
@@ -52,7 +52,7 @@ func resourceArmServiceBusTopicAuthorizationRule() *schema.Resource {
 				Type:         schema.TypeString,
 				Required:     true,
 				ForceNew:     true,
-				ValidateFunc: azure.ValidateServiceBusTopicName(),
+				ValidateFunc: validate.TopicName(),
 			},
 
 			"resource_group_name": azure.SchemaResourceGroupName(),
